lexer: convert re.UNICODE to int
Since python 3.6, RegexFlags are instances of RegexFlag, not int.
Passing re.UNICODE directly to the lexer will result in the generated
lextab.py returning a SyntaxError.

RegexFlag instances need to be explicitely converted to int. This
solution does not break compatibility with previous previous versions of
python that already return int values.

fix #35
ref: https://docs.python.org/3/library/re.html#module-contents